fix: use persistence bech prefix
